migrate to, fix display version and logic
The indication of the problem was that when you used a to version the
version was displayed as the first or the last version of your
migrations, whatever version you were migrating to.

The cause of the issue was reuse of $migrationVersion which was being
set anyway by the foreach loop.

Because of this we also discovered that the comparation during up was
actually wrong, but it did work before because of the display numer
issue.

Signed-off-by: Ike Devolder <ike.devolder@studioemma.eu>