offers install option only when pimcore is not installed
The previous check would check if this plugin was enabled to show the
acutal commands, when not enabled it would only offer to install pimcore
from scratch.

We can actually install the plugin on an existing project. We had to
enable the plugin manually, this is cumbersome and kind of useless.

When the plugin is there and you call the cli script it will enable
itself to allow immediate operation.

Signed-off-by: Ike Devolder <ike.devolder@studioemma.eu>